+++ pkg/aws/v1alpha3/cluster/mutate_cluster_infraref_update_test.go
@@ -64,6 +64,7 @@ func TestMutateInfraRefUpdate(t *testing.T) {
 
 			// try to create the cluster
 			cluster := unittest.DefaultCluster()
+			cluster.Spec.InfrastructureRef.Namespace = ""
 			err = fakeK8sClient.CtrlClient().Create(tc.ctx, cluster)
 			if err != nil {
 				t.Fatal(err)